Issue #36 - Fixed the editor css overflow in the product description page.
And also fixed the larger problem of people not realising that they have the power to
configure tinymce.

I have also removed the width field because it automatically sets itself to the width of
the text area that it is replacing. This is a really good default and should be the one
that we use. People can customise this later if they choose to.